requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
bidict==0.22.1
black==23.3.0
certifi==2022.12.7
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Django==4.2
frozenlist==1.3.3
idna==3.4
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
python-engineio==4.4.0
python-socketio==5.8.0
redis==4.5.4
requests==2.28.2
sqlparse==0.4.3
tzdata==2023.3
urllib3==1.26.15
websocket-client==1.5.1
yarl==1.8.2
